Update counter app with react components
Submit assignment2-1 as the fourth try
