+++ src/Page.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import Counter from './Counter';
 import Buttons from './Buttons';
