+++ src/index.jsx
@@ -1,26 +1,5 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import Page from './Page';
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick(i) {
-    setState({
-      count: count + i,
-    });
-  }
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-    />
-  );
-}
-
+import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
